Handle u64 in napi

+++ tensorzero-core/src/db/datasets.rs
@@ -322,7 +322,7 @@ pub struct GetDatapointsParams {
 #[cfg_attr(test, automock)]
 pub trait DatasetQueries {
     /// Counts rows for a dataset based on query parameters
-    async fn count_rows_for_dataset(&self, params: &DatasetQueryParams) -> Result<u32, Error>;
+    async fn count_rows_for_dataset(&self, params: &DatasetQueryParams) -> Result<u64, Error>;
 
     /// Inserts rows into a dataset table by selecting from the inference tables
     /// Returns the number of rows inserted
